File existens check

+++ encryptor.rb
@@ -40,23 +40,28 @@ class Encryptor
   def encrypt(file, out_name)
     # creates output folder if it is not available
     FileUtils.mkdir('output') if !Dir.exist?('output')
-    file = File.read(file)
-    ln = 0
-    file.each_line{|l|ln += 1}
-    bar = ProgressBar.new(ln)
-    file.each_line do |line|
-      string = ""
-      line.strip.each_char do |char|
-        if char =~ /[A-Z]/
-          letter_case = capital_letters
-        elsif char =~ /[a-z]/
-          letter_case = small_letters
+    if !File.exist?(file)
+      puts "Sorry could not find the file you entered !".colorize(:red)
+      exit
+    else
+      file = File.read(file)
+      ln = 0
+      file.each_line{|l|ln += 1}
+      bar = ProgressBar.new(ln)
+      file.each_line do |line|
+        string = ""
+        line.strip.each_char do |char|
+          if char =~ /[A-Z]/
+            letter_case = capital_letters
+          elsif char =~ /[a-z]/
+            letter_case = small_letters
+          end
+          char =~ /[A-Za-z]/ ? string << letter_case[(letter_case.index(char) + @code) % letter_case.length].to_s : string << char
         end
-        char =~ /[A-Za-z]/ ? string << letter_case[(letter_case.index(char) + @code) % letter_case.length].to_s : string << char
+        File.open("./output/#{out_name}", 'a'){|f| f.puts "#{string}"}
+        sleep 0.1
+        bar.increment!
       end
-      File.open("./output/#{out_name}", 'a'){|f| f.puts "#{string}"}
-      sleep 0.1
-      bar.increment!
     end
   end
   # decryptes a file
